refactor: queries 폴더 index 내보내기 적용
