+++ src/main.tsx
@@ -5,7 +5,7 @@ import { ToastContainer } from "react-toastify";
 import "./styles/globals.css";
 import { App } from "./App.tsx";
 import { RootErrorBoundary } from "./components/error-boundary";
-import { QueryProvider } from "./queries/provider.tsx";
+import { QueryProvider } from "./queries";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
